Cast 0 to return type

+++ tests/data/gmock_test/test_add_function_get/mockup.cc
@@ -8,5 +8,5 @@ int a_get_y2(){
     if(0 != mockup_global_ptr)
         return mockup_global_ptr->a_get_y2();
     else
-        return 0;
+        return (int)0;
 } /* a_get_y2 */
